refactor 자동차 경주(리팩토링)
 - view package 분리
